Reference

ProfileSVG.@profview — Macro

@profview f(args...)

Clear the Profile buffer, profile f(args...), and view the result graphically.

ProfileSVG.view — Function

ProfileSVG.view([fcolor,] data=Profile.fetch(); kwargs...)
ProfileSVG.view([fcolor,] g::FlameGraph; kwargs...)

View profiling results.

keywords for SVG style

bgcolor (default: :fcolor)

The background style. One of :fcolor/:classic/:transparent.

fontcolor (default: :fcolor)

The font color style. One of :fcolor/:classic/:currentcolor/:bw.

frameopacity (default: 1)

The opacity of frames in [0, 1].

yflip (default: false)

If true, the "icicle" graph will be rendered.

align (default: :root)

The alignment of the flame graph. One of :root/:tip.

mindepth (default: 1)

The minimum level of the rows to be rendered.

maxdepth (default: 50)

The maximum level of the rows to be rendered. The maximum number of rows is maxdepth - mindepth + 1.

maxframes (default: 2000)

The maximum number of the rendered frames.

width (default: 960)

The width of output SVG image in pixels.

height (default: 0)

The height of output SVG image in pixels. If you set the height to 0, it will be calculated automatically according to the graph.

roundradius (default: 2)

The rounding radius of the corners of each frame in pixels.

font (default: "inherit")

The font family names for texts. This setting is used as the CSS font-family property, i.e. you can use a comma-separated list.

fontsize (default: 12)

The font size of texts for function information, in pixels (not points).

notext (default: false)

If true, the texts overlaid on the frames will be hidden by the interactive feature.

timeunit (default: :none)

If :s, :ms, :us, or :µs is specified, the duration of the block will be displayed in the lower right corner in the specified unit.

delay (default: 0.0)

The delay between backtraces, in seconds. If a non-positive number is specified, the current setting in Profile.init() will be used.

title (default: "Profile results")

The title (caption) of the graph.

keywords for flamegraph

lidict

C

combine

recur

pruned

filter

See FlameGraphs for details.

ProfileSVG.save — Function

ProfileSVG.save([fcolor,] io::IO, data=Profile.fetch(); kwargs...)
ProfileSVG.save([fcolor,] io::IO, g::FlameGraph; kwargs...)
ProfileSVG.save([fcolor,] filename, data=Profile.fetch(); kwargs...)
ProfileSVG.save([fcolor,] filename, g::FlameGraph; kwargs...)

Save profile results as an SVG file. See ProfileSVG.view for details of kwargs.

ProfileSVG.set_default — Function

ProfileSVG.set_default([fcolor]; kwargs...)

Set defult configurations of profiling results. See ProfileSVG.view for details of kwargs.

Examples

ProfileSVG.set_default(width=600, fontsize=9)

ProfileSVG.init — Function

ProfileSVG.init()

Initialize the settings.